+++ onnxscript/function_libs/torchlib/ops/core.py
@@ -14,8 +14,8 @@ from __future__ import annotations
 from typing import Any, Optional, Sequence, Tuple, Union
 
 from onnxscript import BOOL, DOUBLE, FLOAT, INT8, INT16, INT32, INT64
-from onnxscript.function_libs.torch_aten.registration import torch_op
-from onnxscript.function_libs.torch_aten.tensor_typing import (
+from onnxscript.function_libs.torchlib.registration import torch_op
+from onnxscript.function_libs.torchlib.tensor_typing import (
     IntType,
     RealType,
     TFloat,
