+++ onnxscript/tests/function_libs/torch_aten/ops_correctness_test.py
@@ -44,10 +44,10 @@ from torch.utils import _pytree as pytree
 
 import onnxscript
 import onnxscript.evaluator
-from onnxscript.function_libs.torch_aten import graph_building
-from onnxscript.function_libs.torch_aten.ops import core as core_ops
-from onnxscript.function_libs.torch_aten.ops import nn as nn_ops
-from onnxscript.function_libs.torch_aten.ops import special as special_ops
+from onnxscript.function_libs.torchlib import graph_building
+from onnxscript.function_libs.torchlib.ops import core as core_ops
+from onnxscript.function_libs.torchlib.ops import nn as nn_ops
+from onnxscript.function_libs.torchlib.ops import special as special_ops
 from onnxscript.tests.common import version_utils
 from onnxscript.tests.function_libs.torch_aten import extra_opinfo
